removing initial implementation
