+++ tbsim/interventions/__init__.py
@@ -7,15 +7,14 @@ This module contains all intervention classes for the TB simulation.
 # Import all intervention classes
 from .interventions import TBProductRoutine
 from .beta import *
-from .tpt import TPTTx, TPTSimple, TPTHousehold
-from .tpt2 import (
+from .tpt import (
     RegimenCategory,
     TPTRegimen,
     REGIMENS,
     TPTProduct,
     TPTDelivery,
     TPTRoutine,
-    TPTHousehold as TPTHouseholdV2,
+    TPTHousehold,
 )
 from .bcg import BCGVx, BCGRoutine
 from .tb_treatment import TBTreatment, EnhancedTBTreatment, create_dots_treatment, create_dots_improved_treatment, create_first_line_treatment
@@ -27,16 +26,13 @@ from .tb_diagnostic import TBDiagnostic, EnhancedTBDiagnostic
 __all__ = [
     'TBProductRoutine',
     'BetaByYear',
-    'TPTTx',
-    'TPTSimple',
-    'TPTHousehold',
     'RegimenCategory',
     'TPTRegimen',
     'REGIMENS',
     'TPTProduct',
     'TPTDelivery',
     'TPTRoutine',
-    'TPTHouseholdV2',
+    'TPTHousehold',
     'BCGVx',
     'BCGRoutine',
     'TBTreatment',
